etreco_app/requirements.txt
beautifulsoup4==4.12.3
black==24.4.2
bsedata==0.6.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
frozendict==2.4.4
fuzzywuzzy==0.18.0
html5lib==1.1
idna==3.7
Levenshtein==0.25.1
lxml==5.2.2
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
peewee==3.17.5
platformdirs==4.2.2
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2024.1
rapidfuzz==3.9.3
requests==2.32.3
schedule==1.2.2
six==1.16.0
soupsieve==2.5
thefuzz==0.22.1
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
wheel==0.43.0
yfinance==0.2.40
